qr_code/requirements.txt
python-dotenv
qrcode[pil]